backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.1.0

# AI & Agents
openai>=1.66.5
openai-agents>=0.0.10
cohere>=5.5.8
google-generativeai>=0.8.5

# Vector Database
qdrant-client>=1.8.0

# Database
psycopg2-binary==2.9.9
sqlalchemy>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.1

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
torch==2.1.1
numpy==1.24.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx>=0.27.0

# Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Rate Limiting
slowapi>=0.1.9